Controller: rotes findById and findByName fixed. Server works

+++ src/main/java/com/sitool/cardsdictionary/card/service/CardService.java
@@ -11,7 +11,7 @@ public interface CardService {
 
     Boolean addCard(CardAddDto card);
 
-    CardDto getCardById(Integer cardId);
+    CardDto getCardById(Long cardId);
 
     CardDto getCardByName(String name);
